+++ saleor/order/tests/webhooks/test_exclude_shipping_cache.py
@@ -10,10 +10,10 @@ from prices import Money
 
 from ....core.prices import quantize_price
 from ....shipping.interface import ShippingMethodData
+from ....shipping.webhooks.shared import CACHE_EXCLUDED_SHIPPING_TIME
 from ....webhook.event_types import WebhookEventSyncType
 from ....webhook.transport.utils import generate_cache_key_for_webhook
 from ...webhooks.exclude_shipping import (
-    CACHE_EXCLUDED_SHIPPING_TIME,
     excluded_shipping_methods_for_order,
 )
